# storage/api-docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Knygų rekomendacijos API",
        "description": "API darbui su kategorijomis, knygomis ir rekomendacijomis",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/knygos": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "operationId": "cbc0ea74f2f6a8d38ceece8bf3c9ca7c",
                "responses": {
                    "200": {
                        "description": "Successful request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Creates a new book. The `id` is auto-generated and should not be provided.",
                "operationId": "1e8416a5e17d2c079ab6d2db020d2c81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pavadinimas",
                                    "autorius",
                                    "isbn",
                                    "kategorija_id"
                                ],
                                "properties": {
                                    "pavadinimas": {
                                        "type": "string",
                                        "example": "Nauja knyga"
                                    },
                                    "autorius": {
                                        "type": "string",
                                        "example": "Autorius Vardas"
                                    },
                                    "aprasymas": {
                                        "type": "string",
                                        "example": "Trumpas aprašymas"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-3-16-148410-0"
                                    },
                                    "kategorija_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "pavadinimas": {
                                            "type": "string",
                                            "example": "Nauja knyga"
                                        },
                                        "autorius": {
                                            "type": "string",
                                            "example": "Autorius Vardas"
                                        },
                                        "aprasymas": {
                                            "type": "string",
                                            "example": "Trumpas aprašymas"
                                        },
                                        "isbn": {
                                            "type": "string",
                                            "example": "978-3-16-148410-0"
                                        },
                                        "kategorija_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The pavadinimas field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/knygos/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a single book",
                "operationId": "b3f027d4b0b48e7b357d11a8c03cfd1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update a book",
                "description": "Update all fields of a book. The `id` is auto-generated and should not be changed.",
                "operationId": "1d17e14837df0d7789579b68e05f90e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pavadinimas",
                                    "autorius",
                                    "isbn",
                                    "kategorija_id"
                                ],
                                "properties": {
                                    "pavadinimas": {
                                        "type": "string",
                                        "example": "Atnaujinta knyga"
                                    },
                                    "autorius": {
                                        "type": "string",
                                        "example": "Autorius Vardas"
                                    },
                                    "aprasymas": {
                                        "type": "string",
                                        "example": "Atnaujintas aprašymas"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-3-16-148410-0"
                                    },
                                    "kategorija_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Knyga"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book (owner only)",
                "operationId": "e623648d2ffc040750f815f4b8479c1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Has recommendations"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Books"
                ],
                "summary": "Partially update a book",
                "description": "Update one or more fields of a book. The `id` is auto-generated.",
                "operationId": "9cef7fa02450358b43e715d69673cc91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pavadinimas": {
                                        "type": "string",
                                        "example": "Atnaujinta knyga"
                                    },
                                    "autorius": {
                                        "type": "string",
                                        "example": "Autorius Vardas"
                                    },
                                    "aprasymas": {
                                        "type": "string",
                                        "example": "Atnaujintas aprašymas"
                                    },
                                    "isbn": {
                                        "type": "string",
                                        "example": "978-3-16-148410-0"
                                    },
                                    "kategorija_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Knyga"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kategorijos/{kategorijaId}/knygos/{knygaId}/rekomendacijos": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get recommendations for a book in a category",
                "description": "Retrieve all recommendations for a specific book within a specific category.",
                "operationId": "439c07bd0e205b4ee48b38bb282d60cd",
                "parameters": [
                    {
                        "name": "kategorijaId",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "knygaId",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recommendations or a message if none exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Rekomendacija"
                                            }
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Šiai knygai dar nėra rekomendacijų"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category or book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Knyga nerasta šioje kategorijoje"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/knygos/{id}/admin-delete": {
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Admin deletes a book",
                "operationId": "b4f9a6d8f18e8fc29abd9df76514176c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted by admin"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Knyga": {
                "required": [
                    "pavadinimas",
                    "autorius",
                    "isbn",
                    "kategorija_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pavadinimas": {
                        "type": "string",
                        "example": "Nauja knyga"
                    },
                    "autorius": {
                        "type": "string",
                        "example": "Autorius Vardas"
                    },
                    "aprasymas": {
                        "type": "string",
                        "example": "Trumpas aprašymas"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "978-3-16-148410-0"
                    },
                    "kategorija_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "kategorija": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "pavadinimas": {
                                "type": "string",
                                "example": "Fantastika"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}